requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-socketio==5.11.0
python-dotenv==1.0.1
pymongo==4.6.1
motor==3.3.2
dnspython==2.5.0
aiortc==1.5.0
aiohttp==3.9.3
python-multipart==0.0.9
websockets==12.0
